Small UNSAFE script to check the examples.
Be careful when running running the script! It assumes everything is
syntactically correct, and doesn't check the values passed to the
terminal!

This script was made to test the examples on my system. It assumes two
repositories: one on the branch of the paper, and another on the branch
of the code. Also, I'm using stack in order to use a recent GHC and
a stable set of libraries.

It shouldn't be too hard to modify it to run on other systems, but this
isn't my goal at the moment. It was quickly developed to check if I
didn't make any mistake when writing the examples.